Set up the PIC and interrupts for the keyboard
The translation of scancodes into characters is done via a
third-party library. Additionally to show the keyboard in action, I
improved the VGA text mode to support newlines, scrolling and cursor
movement. Other than that, I did a few cleanups and wrapped global
variables in mutexes based on spinning.

+++ src/libstd.rs
@@ -55,10 +55,10 @@ pub unsafe extern fn memmove(dest: *mut u8, src: *const u8,
 }
 
 #[no_mangle]
-pub unsafe extern fn memset(s: *mut u8, c: i32, n: usize) -> *mut u8 {
+pub unsafe extern fn memset(s: *mut u8, c: u8, n: usize) -> *mut u8 {
     let mut i = 0;
     while i < n {
-        *s.offset(i as isize) = c as u8;
+        *s.offset(i as isize) = c;
         i += 1;
     }
     return s;
